Update to projectile animation & Cleanup of code
Cleaned(*) up Game.cpp code, still more work to do but at least the init and restart is much cleaner.
also updated the projectile animation right now i fixes the position and animation of it.
but only shows a  texture for the first bullet still need to fix that up

side note: updated readme a  bit.

*with cleaned the code i meant i sent it to respective cpps so it would not be a bunch of code in the game init or update.