update to menu version display, buttons for game & hp value draw
updated the hp value draw to be placed inside the hp bar so u have a  visual and numeral value of how much hp left you have.
i added the return to game button on the pause menu so it makes a  cleaner loop with the pause menu.
added to the player collision mechanic with slime for the player to be placed in the center of the screen and also changed the damage values of the collisions for the different slimes.
added a display of current version to the menu, the version is stored in global.cpp